Add Hibernate
* changed layer repo from jdbc to hibernate
* fixed minor bugs

+++ src/main/java/ru/effective_mobile/todo/model/PaginatedResponse.java
@@ -8,12 +8,12 @@ import java.util.List;
 public class PaginatedResponse<T> {
 
     private final List<T> items;
-    private final int totalElements;
+    private final long totalElements;
     private final int totalPages;
     private final int currentPage;
     private final int pageSize;
 
-    public PaginatedResponse(List<T> items, int totalElements, int currentPage, int pageSize) {
+    public PaginatedResponse(List<T> items, long totalElements, int currentPage, int pageSize) {
         this.items = items;
         this.totalElements = totalElements;
         this.totalPages = (int) Math.ceil((double) totalElements / pageSize);
